feat: Fix name and add release and dev profiles
So the idea is to not use incremental and only one codegen-unit so that
binaries will be the same  if there are no changes.

During dev we up do use incremental and up codegen-units as repeatable
binaries isn't necessary?

Also, I added ~/.cargo/config.toml and use sccache which speeds up
builds significantly after the cache is warm. A release build changes
from 42s to 11s and debug builds from 20 to 10.