feat: Add panic = "abort" remove build.rs
This save 2.3MB and drops the box-with-tri-hole to 9.6MB. Also
explicitly added debug, lto and strip to be explicit.

Remove build.rs I had added it as I discussed with Claude,
https://claude.ai/share/7b90d2a7-db84-475b-8359-63e3bd4aaa5a,
how to make reproducible builds. At the time was worried about
timestamps changing in the binary and thus binaries not be exactly
reproducible. Turns out there in none om the box-with-tri-hole
binary so build.rs is unnecessary.

I've move it to scripts in case I do need it.